backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
pandas
duckdb
